Fix MSVC compilation problem
git-svn-id: svn://svn.code.sf.net/p/octave/code/trunk/octave-forge@5076 fc592f45-0881-4044-91ff-c96495a54512

+++ main/signal/src/upfirdn.cc
@@ -90,7 +90,7 @@ Upsample, FIR filtering and downsample.@*\n\
 
   double r=p/((double) q);
 
-  int Ly= ceil( ((Lx-1)*p + Lh)/q );
+  int Ly= ceil( ((Lx-1)*p + Lh)/((double) q) );
 
   Matrix y(Ly,cx,0.0);
